refactor(proteus): split proteus.rs into topic-based submodules
Split the 848-line crypto/src/proteus.rs into crypto/src/proteus/ with
files grouped by concern (conversation_session, session, message, prekey,
core_crypto), co-locating tests with the code they exercise. Public API
unchanged: ProteusCentral, ProteusConversationSession, and SessionIdentifier
remain reachable at crate::proteus::* via mod.rs re-exports. Prerequisite
for the upcoming group_store refactor.
